mcp/pyproject.toml
[project]
name = "finpy-mcp"
version = "0.1.0"
description = "MCP server for querying finpy silver tables"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "trino>=0.330.0",
    "pyarrow>=15.0.0",
    "python-dotenv>=1.0.0",
    "pendulum>=3.0.0",
]

[project.scripts]
finpy-mcp = "server:main"
